DND Improvements in OO and Efficiency
