+++ jhotdraw6/src/org/jhotdraw/contrib/dnd/DNDHelper.java
@@ -27,22 +27,56 @@ import CH.ifa.draw.util.*;
  * @author  C.L.Gilbert <dnoyeb@users.sourceforge.net>
  * @version <$CURRENT_VERSION$>
  */
-public abstract class DNDHelper implements DragGestureListener {
+public abstract class DNDHelper {
 	public static DataFlavor ASCIIFlavor = new DataFlavor("text/plain; charset=ascii", "ASCII text");
 	private DragGestureRecognizer dgr;
-	private DropTarget dt;
-	private Boolean autoscrollState;
+	private DragGestureListener dragGestureListener;
+	private DropTarget dropTarget;
 	private DragSourceListener dragSourceListener;
 	private DropTargetListener dropTargetListener;
 	
+	public DNDHelper(){
+	}
+	/**
+	 * Do not call this from the constructor.  its methods are overridable.
+	 */
+	public void initialize(DragGestureListener dgl) {
+		if(isDragSource()) {
+			setDragGestureListener( dgl );
+			setDragSourceListener( createDragSourceListener() );
+			setDragGestureRecognizer(createDragGestureRecognizer(getDragGestureListener()));
+		}
+		if(isDropTarget()) {
+			setDropTargetListener( createDropTargetListener() );
+			setDropTarget(createDropTarget());
+		}
+	}
+	public void deinitialize(){
+		if(getDragSourceListener() != null){
+			destroyDragGestreRecognizer();
+			setDragSourceListener( null );
+		}
+		if(getDropTargetListener() != null){
+			setDropTarget( null );
+			setDropTargetListener( null );
+		}
+	}
+	public void setDragSourceState(boolean state) {
+		if(state == false){
+			getDragGestureRecognizer().setSourceActions(DnDConstants.ACTION_NONE);
+		}
+		else {
+			getDragGestureRecognizer().setSourceActions(getDragSourceActions());
+		}
+	}
 	abstract protected DrawingView view();
 	abstract protected DrawingEditor editor();
+	abstract protected boolean isDragSource();
+	abstract protected boolean isDropTarget();
 
-	public DNDHelper(){
-		setDragSourceListener( createDragSourceListener() );
-		setDropTargetListener( createDropTargetListener() );
-	}
-
+	
+	
+	
 	protected static Object ProcessReceivedData(DataFlavor flavor, Transferable transferable) {
 		if (transferable == null) {
 			return null;
@@ -101,33 +135,34 @@ public abstract class DNDHelper implements DragGestureListener {
 			return null;
 		}
 	}
-
-	public boolean setDropTargetActive(boolean state) {
-		if (state == true) {
-		    return createDropTarget();
-		}
-		else {
-			destroyDropTarget();
-			return false;
-		}
+	protected int getDragSourceActions() {
+		return DnDConstants.ACTION_COPY_OR_MOVE;
 	}
-
-	public boolean setDragSourceActive(boolean state) {
-		if (state == true) {
-			return createDragGestureRecognizer();
-		}
-		else {
-			destroyDragGestreRecognizer();
-			return false;
-		}
+	protected void setDragGestureListener(DragGestureListener dragGestureListener){
+		this.dragGestureListener = dragGestureListener;
 	}
-
-	protected boolean createDropTarget() {
-		if (Component.class.isInstance(view())) {
+	protected DragGestureListener getDragGestureListener(){
+		return dragGestureListener;
+	}
+	protected void setDragGestureRecognizer(DragGestureRecognizer dragGestureRecognizer){
+		dgr = dragGestureRecognizer;
+	}
+	protected DragGestureRecognizer getDragGestureRecognizer(){
+		return dgr;
+	}
+	protected void setDropTarget(DropTarget dropTarget){
+		if((dropTarget == null) && (this.dropTarget != null)){
+			this.dropTarget.setComponent(null);
+			this.dropTarget.removeDropTargetListener( getDropTargetListener() );
+		}
+		this.dropTarget = dropTarget;
+	}
+	protected DropTarget createDropTarget() {
+		DropTarget dt = null;
+		if (view() instanceof Component) {
 			try {
 				dt = new DropTarget((Component)view(), DnDConstants.ACTION_COPY_OR_MOVE, getDropTargetListener());
-				//System.out.println( "" + view() + " Initialized to DND.");
-				return true;
+				System.out.println(view().toString() + " Initialized to DND.");
 			}
 			catch (java.lang.NullPointerException npe) {
 				System.err.println("View Failed to initialize to DND.");
@@ -136,550 +171,55 @@ public abstract class DNDHelper implements DragGestureListener {
 				npe.printStackTrace();
 			}
 		}
-		return false;
+		return dt;
 	}
-	protected void destroyDropTarget() {
-		if (dt!= null) {
-			dt.setComponent(null);
-			dt.removeDropTargetListener( getDropTargetListener() );
-			dt = null;
-		}
-	}
-
 
 	/**
 	 * Used to create the gesture recognizer which in effect turns on draggability.
 	 */
-	protected boolean createDragGestureRecognizer() {
-		if (Component.class.isInstance(view())) {
+	protected DragGestureRecognizer createDragGestureRecognizer(DragGestureListener dgl) {
+		DragGestureRecognizer aDgr = null;
+		if (view() instanceof Component) {
 			Component c = (Component)view();
-			dgr =	DragSource.getDefaultDragSource().createDefaultDragGestureRecognizer(
+			aDgr =	DragSource.getDefaultDragSource().createDefaultDragGestureRecognizer(
 					c,
-					DnDConstants.ACTION_COPY_OR_MOVE,
-					this);
-			//System.out.println("DragGestureRecognizer created: " + dgl);
-			return true;
+					getDragSourceActions(),
+					dgl);
+			System.out.println("DragGestureRecognizer created: " + view());
 		}
-		else
-			return false;
+		return aDgr;
 	}
 
 	/**
 	 * Used to destroy the gesture listener which in effect turns off dragability.
 	 */
 	protected void destroyDragGestreRecognizer() {
-		//System.out.println("Destroying DGR");
-		if (dgr != null) {
-			dgr.removeDragGestureListener(this);
-	    	dgr.setComponent(null);
-			dgr = null;
+		System.out.println("Destroying DGR " + view());
+		if (getDragGestureRecognizer() != null) {
+			getDragGestureRecognizer().removeDragGestureListener(getDragGestureListener());
+	    	getDragGestureRecognizer().setComponent(null);
+			setDragGestureRecognizer(null);
 		}
 	}
 
-	/*******************************************DragGestureListener*****************/
-
-	/**
-	 * This function is called when the drag action is detected.  If it agrees
-	 * with the attempt to drag it calls startDrag(), if not it does nothing.
-	 */
-	public void dragGestureRecognized(DragGestureEvent dge) {
-		Component c = dge.getComponent();
-		//System.out.println("recognized for " + c);
-
-		if (DrawingView.class.isInstance(c)) {
-			boolean found = false;
-			DrawingView dv = (DrawingView)c;
-			/* Send the drawing view which inspired the action a mouseUp to clean
-			up its current tool.  THis is because mouse up will otherwise never
-			be send and the tool will be stuck with only mouse down which means
-			it will likely stay activated.  solve later for not just make
-			but report. */
-			/* this is a list of cloned figures */
-			FigureEnumeration selectedElements = dv.selection();
-
-			if (selectedElements.hasNextFigure() == false) {
-				return;
-			}
-
-			Point p = dge.getDragOrigin();
-//				System.out.println("origin at " + p);
-			while (selectedElements.hasNextFigure()) {
-				Figure f = selectedElements.nextFigure();
-				if (f.containsPoint(p.x, p.y)) {
-/*              Rectangle r = figgy.displayBox();
-					sx = r.width;
-					sy = r.height;*/
-					//System.out.println("figure is " + figgy);
-					found = true;
-					break;
-				}
-			}
-			if (found == true) {
-				DNDFigures dndff = new DNDFigures(dv.selection(), p);
-				DNDFiguresTransferable trans = new DNDFiguresTransferable(dndff);
-
-				/* SAVE FOR FUTURE DRAG IMAGE SUPPORT */
-				/* drag image support that I need to test on some supporting platform.
-				windows is not supporting this on NT so far. Ill test 98 and 2K next
-
-				boolean support = dragSource.isDragImageSupported();
-				java.awt.image.BufferedImage  bi = new BufferedImage(sx,sy,BufferedImage.TYPE_INT_RGB);
-				Graphics2D g = bi.createGraphics();
-				Iterator itr2 = selectedElements.iterator();
-				while ( itr2.hasNext() ) {
-					Figure fig = (Figure) itr2.next();
-					fig = (Figure)fig.clone();
-					Rectangle rold = fig.displayBox();
-					fig.moveBy(-rold.x,-rold.y);
-					fig.draw(g);
-				}
-				g.setBackground(Color.red);
-				dge.getDragSource().startDrag(
-								dge,
-								DragSource.DefaultMoveDrop,
-								bi,
-								new Point(0,0),
-								trans,
-								this);
-				*/
-				if (JComponent.class.isInstance( c )) {
-				    ((JComponent)c).setAutoscrolls(false);
-				}
-				dge.getDragSource().startDrag(
-								dge,
-								null,
-								trans,
-								getDragSourceListener());
-			}
-		}
-	}
-	/***************************************End DragGestureListener*****************/
-
 	
-	
-	
-	
-	
-	
-	
-	
-	private Undoable targetUndoable;
-	/**
-	 * Factory method for undo activity
-	 */
-	protected Undoable createTargetUndoActivity(DrawingView view) {
-		return new AddUndoActivity( view );
-	}
-	protected void setTargetUndoActivity(Undoable undoable){
-		targetUndoable = undoable;
-	}
-	protected Undoable getTargetUndoActivity(){
-		return targetUndoable;
-	}
-	public static class AddUndoActivity extends UndoableAdapter {
-
-		public AddUndoActivity(DrawingView newDrawingView) {
-			super(newDrawingView);
-			//System.out.println("AddUndoActivity created " + newDrawingView);			
-			setUndoable(true);
-			setRedoable(true);
-		}
-
-		public boolean undo() {
-			if (!super.undo()) {
-				return false;
-			}
-			//undo of add really shouldnt need visitor !?!dnoyeb!?!
-			//System.out.println("AddUndoActivity AddUndoActivity undo");
-			DeleteFromDrawingVisitor deleteVisitor = new DeleteFromDrawingVisitor(getDrawingView().drawing());
-			FigureEnumeration fe = getAffectedFigures();
-			while (fe.hasNextFigure()) {
-	    		Figure f = fe.nextFigure();
-				f.visit(deleteVisitor);
-			}
-			setAffectedFigures( deleteVisitor.getDeletedFigures() );
-			getDrawingView().clearSelection();
-			return true;
-		}
-
-		public boolean redo() {
-			// do not call execute directly as the selection might has changed
-			if (!isRedoable()) {
-				return false;
-			}
-			//System.out.println("AddUndoActivity redo");
-			getDrawingView().clearSelection();
-			setAffectedFigures(getDrawingView().insertFigures(
-				getAffectedFigures(), 0, 0, false));
-
-			return true;
-		}
-	}
-
-	/*******************************************DropTargetListener*****************/
-	
-	
-
-
-	private void setDropTargetListener(DropTargetListener dropTargetListener){
+	protected void setDropTargetListener(DropTargetListener dropTargetListener){
 		this.dropTargetListener = dropTargetListener;
 	}
-	private DropTargetListener getDropTargetListener(){
+	protected DropTargetListener getDropTargetListener(){
 		return dropTargetListener;
 	}
-	private DropTargetListener createDropTargetListener(){
-		return new innerDropTargetListener();
+	protected DropTargetListener createDropTargetListener(){
+		return new JHDDropTargetListener(editor(),view());
 	}
-
-	private class innerDropTargetListener implements DropTargetListener {
-		private int     fLastX=0, fLastY=0;      // previous mouse position
-		/**
-		 * Called when a drag operation has encountered the DropTarget.
-		 */
-		public void dragEnter(DropTargetDragEvent dtde) {
-			//System.out.println("DropTargetDragEvent-dragEnter");
-			supportDropTargetDragEvent(dtde);
-			if (fLastX == 0) {
-				fLastX = dtde.getLocation().x;
-			}
-			if (fLastY == 0) {
-				fLastY = dtde.getLocation().y;
-			}
-		}
-
-		/**
-		 * The drag operation has departed the DropTarget without dropping.
-		 */
-		public void dragExit(DropTargetEvent dte) {
-			//System.out.println("DropTargetEvent-dragExit");
-		}
-
-		/**
-		 * Called when a drag operation is ongoing on the DropTarget.
-		 */
-		 public void dragOver(DropTargetDragEvent dtde) {
-			//System.out.println("DropTargetDragEvent-dragOver");
-			if (supportDropTargetDragEvent(dtde)==true) {
-				int x=dtde.getLocation().x;
-				int y=dtde.getLocation().y;
-				if ((Math.abs(x - fLastX) > 0) || (Math.abs(y - fLastY) > 0) ) {
-					//FigureEnumeration fe = view().selectionElements();
-					//while (fe.hasNextFigure()) {
-					//	fe.nextFigure().moveBy(x - fLastX, y - fLastY);
-					//	System.out.println("moving Figures " + view());
-					//}
-					//view().drawing().update();
-					fLastX = x;
-					fLastY = y;
-				}
-			}
-		 }
-
-		/**
-		 * The drag operation has terminated with a drop on this DropTarget.
-		 * Be nice to somehow incorporate FigureTransferCommand here.
-		 */
-		public void drop(DropTargetDropEvent dtde) {
-			//System.out.println("DropTargetDropEvent-drop");
-
-			if (dtde.isDataFlavorSupported(DNDFiguresTransferable.DNDFiguresFlavor) == true) {
-				//System.out.println("DNDFiguresFlavor");
-				if ((dtde.getDropAction() & DnDConstants.ACTION_COPY_OR_MOVE) != 0 ) {
-					//System.out.println("copy or move");
-					if (dtde.isLocalTransfer() == false) {
-						System.err.println("Intra-JVM Transfers not implemented for figures yet.");
-						return;
-					}
-					dtde.acceptDrop(dtde.getDropAction());
-					try { /* protection from a malicious dropped object */
-						setTargetUndoActivity( createTargetUndoActivity( view() ) );
-						DNDFigures ff = (DNDFigures)ProcessReceivedData(DNDFiguresTransferable.DNDFiguresFlavor, dtde.getTransferable());
-						getTargetUndoActivity().setAffectedFigures( ff.getFigures() );
-						Point theO = ff.getOrigin();
-						view().clearSelection();
-						Point newP = dtde.getLocation();
-						/** origin is where the figure thinks it is now
-						  * newP is where the mouse is now.
-						  * we move the figure to where the mouse is with this equation
-						  */
-						int dx = newP.x - theO.x;  /* distance the mouse has moved */
-						int dy = newP.y - theO.y;  /* distance the mouse has moved */
-						//System.out.println("mouse at " + newP);
-						FigureEnumeration fe = view().insertFigures( getTargetUndoActivity().getAffectedFigures() ,  dx, dy, false );
-						getTargetUndoActivity().setAffectedFigures( fe );
-
-						if (dtde.getDropAction() == DnDConstants.ACTION_MOVE)
-							view().addToSelectionAll( getTargetUndoActivity().getAffectedFigures() );
-
-						view().drawing().update();
-						editor().getUndoManager().pushUndo( getTargetUndoActivity() );
-						editor().getUndoManager().clearRedos();
-						// update menus
-						editor().figureSelectionChanged( view() );
-						dtde.getDropTargetContext().dropComplete(true);
-					}
-					catch (NullPointerException npe) {
-						npe.printStackTrace();
-						dtde.getDropTargetContext().dropComplete(false);
-					}
-				}
-				else {
-					dtde.rejectDrop();
-				}
-			}
-			else if (dtde.isDataFlavorSupported(DataFlavor.stringFlavor)) {
-				//System.out.println("String flavor dropped.");
-				dtde.acceptDrop(dtde.getDropAction());
-				Object o = ProcessReceivedData(DataFlavor.stringFlavor, dtde.getTransferable());
-				if (o != null) {
-					//System.out.println("Received string flavored data.");
-					dtde.getDropTargetContext().dropComplete(true);
-				}
-				else {
-					dtde.getDropTargetContext().dropComplete(false);
-				}
-			}
-			else if (dtde.isDataFlavorSupported(ASCIIFlavor) == true) {
-				//System.out.println("ASCII Flavor dropped.");
-				dtde.acceptDrop(DnDConstants.ACTION_COPY);
-				Object o = ProcessReceivedData(ASCIIFlavor, dtde.getTransferable());
-				if (o!= null) {
-					//System.out.println("Received ASCII Flavored data.");
-					dtde.getDropTargetContext().dropComplete(true);
-					//System.out.println(o);
-				}
-				else {
-					dtde.getDropTargetContext().dropComplete(false);
-				}
-			}
-			else if (dtde.isDataFlavorSupported(DataFlavor.javaFileListFlavor)) {
-				//System.out.println("Java File List Flavor dropped.");
-				dtde.acceptDrop(DnDConstants.ACTION_COPY);
-				java.io.File [] fList = (java.io.File[]) ProcessReceivedData(DataFlavor.javaFileListFlavor, dtde.getTransferable());
-				if (fList != null) {
-					//System.out.println("Got list of files.");
-					for (int x=0; x< fList.length; x++ ) {
-						System.out.println(fList[x].getAbsolutePath());
-					}
-					dtde.getDropTargetContext().dropComplete(true);
-				}
-				else {
-					dtde.getDropTargetContext().dropComplete(false);
-				}
-			}
-			fLastX = 0;
-			fLastY = 0;
-		}
-
-		/**
-		 * Called if the user has modified the current drop gesture.
-		 */
-		public void dropActionChanged(DropTargetDragEvent dtde) {
-			//System.out.println("DropTargetDragEvent-dropActionChanged");
-			supportDropTargetDragEvent(dtde);
-		}
-	}
-	/**
-	 * Tests wether the Drag event is of a type that we support handling
-	 * Check the DND interface and support the events it says it supports
-	 * if not a dnd interface comp, then dont support! because we dont even
-	 * really know what kind of view it is.
-	 */
-	protected boolean supportDropTargetDragEvent(DropTargetDragEvent dtde) {
-		if (dtde.isDataFlavorSupported(DNDFiguresTransferable.DNDFiguresFlavor) == true) {
-			if (dtde.getDropAction() == DnDConstants.ACTION_COPY) {
-				dtde.acceptDrag(DnDConstants.ACTION_COPY);
-				return true;
-			}
-			else if (dtde.getDropAction() == DnDConstants.ACTION_MOVE) {
-				dtde.acceptDrag(DnDConstants.ACTION_MOVE);
-				return true;
-			}
-			else {
-				dtde.rejectDrag();
-				return false;
-			}
-		}
-		else if (dtde.isDataFlavorSupported(ASCIIFlavor) == true) {
-			dtde.acceptDrag(dtde.getDropAction());
-			return true;
-		}
-		else if (dtde.isDataFlavorSupported(DataFlavor.stringFlavor) == true) {
-			dtde.acceptDrag(dtde.getDropAction());
-			return true;
-		}
-		else if (dtde.isDataFlavorSupported(DataFlavor.javaFileListFlavor) == true) {
-			dtde.acceptDrag(dtde.getDropAction());
-			return true;
-		}
-		else {
-			dtde.rejectDrag();
-			return false;
-		}
-	}
-	/***************************************End DropTargetListener*****************/
-
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	private Undoable sourceUndoable;
-	/**
-	 * Factory method for undo activity
-	 */
-	protected Undoable createSourceUndoActivity(DrawingView drawingView) {
-		return new RemoveUndoActivity( drawingView );
-	}
-	protected void setSourceUndoActivity(Undoable undoable){
-		sourceUndoable = undoable;
-	}
-	protected Undoable getSourceUndoActivity(){
-		return sourceUndoable;
-	}
-	public static class RemoveUndoActivity extends UndoableAdapter {
-		public RemoveUndoActivity(DrawingView view) {
-			super( view );
-			//System.out.println("RemoveUndoActivity created " + view);
-			setUndoable(true);
-			setRedoable(true);
-		}
-
-		public boolean undo() {
-			if (super.undo() && getAffectedFigures().hasNextFigure()) {
-				//System.out.println("RemoveUndoActivity undo");
-				getDrawingView().clearSelection();
-				setAffectedFigures( getDrawingView().insertFigures(
-											getAffectedFigures(), 0, 0,false));
-				return true;
-			}
-			return false;
-		}
-
-		public boolean redo() {
-			// do not call execute directly as the selection might has changed
-			if (isRedoable()) {
-				//System.out.println("RemoveUndoActivity redo");
-				DeleteFromDrawingVisitor deleteVisitor = new DeleteFromDrawingVisitor( getDrawingView().drawing());
-				FigureEnumeration fe = getAffectedFigures();
-				while (fe.hasNextFigure()) {
-					fe.nextFigure().visit(deleteVisitor);
-				}
-				getDrawingView().clearSelection();
-				setAffectedFigures( deleteVisitor.getDeletedFigures() );
-				return true;
-			}
-			return false;
-		}
-		/**
-		 * Releases all resources related to an undoable activity
-		 */
-		public void release() {
-			FigureEnumeration fe = getAffectedFigures();
-			while (fe.hasNextFigure()) {
-				fe.nextFigure().release();
-			}
-			setAffectedFigures(CH.ifa.draw.standard.FigureEnumerator.getEmptyEnumeration());
-		}		
-	}
-	
-	
-	
-	
-	/************************************ Begin DragSourceListener*****************/
-	private DragSourceListener getDragSourceListener(){
+	public DragSourceListener getDragSourceListener(){
 		return dragSourceListener;
 	}
-	private void setDragSourceListener(DragSourceListener dragSourceListener){
+	protected void setDragSourceListener(DragSourceListener dragSourceListener){
 		this.dragSourceListener = dragSourceListener;
 	}
-	private DragSourceListener createDragSourceListener(){
-		return new innerDragSourceListener();
-	}
-	private class innerDragSourceListener implements DragSourceListener {
-		/**
-		 * This method is invoked to signify that the Drag and Drop operation is complete.
-		 * This is the last method called in the process.
-		 */
-		public void dragDropEnd(DragSourceDropEvent dsde) {
-			DrawingView view = (DrawingView) dsde.getDragSourceContext().getComponent();
-			//System.out.println("DragSourceDropEvent-dragDropEnd");
-			if (dsde.getDropSuccess() == true) {
-				if (dsde.getDropAction() == DnDConstants.ACTION_MOVE) {
-	//                System.out.println("DragSourceDropEvent-ACTION_MOVE");
-					//get the flavor in order of ease of use here.
-					setSourceUndoActivity(  createSourceUndoActivity( view ) );
-					DNDFigures df = (DNDFigures)ProcessReceivedData(DNDFiguresTransferable.DNDFiguresFlavor, dsde.getDragSourceContext().getTransferable());
-					getSourceUndoActivity().setAffectedFigures( df.getFigures() );
-
-					//all this visitation needs to be hidden in a view method.
-					DeleteFromDrawingVisitor deleteVisitor = new DeleteFromDrawingVisitor(view.drawing());
-					FigureEnumeration fe = getSourceUndoActivity().getAffectedFigures();
-					while (fe.hasNextFigure()) {
-						fe.nextFigure().visit(deleteVisitor);
-					}
-					view.clearSelection();
-					view.drawing().update();//we made a change to the drawing, so update it.
-
-					editor().getUndoManager().pushUndo( getSourceUndoActivity() );
-					editor().getUndoManager().clearRedos();
-					// update menus
-					editor().figureSelectionChanged( view );
-				}
-				else if (dsde.getDropAction() == DnDConstants.ACTION_COPY) {
-					//System.out.println("DragSourceDropEvent-ACTION_COPY not implemented?");
-				}
-			}
-
-			if (autoscrollState != null) {
-				Component c = dsde.getDragSourceContext().getComponent();
-				if (JComponent.class.isInstance( c )) {
-					JComponent jc = (JComponent)c;
-					jc.setAutoscrolls(autoscrollState.booleanValue());
-					autoscrollState= null;
-				}
-			}
-		}
-		/**
-		 * Called as the hotspot enters a platform dependent drop site.
-		 */
-		public void dragEnter(DragSourceDragEvent dsde) {
-			if (autoscrollState == null) {
-				Component c = dsde.getDragSourceContext().getComponent();
-				if (JComponent.class.isInstance( c )) {
-					JComponent jc = (JComponent)c;
-					autoscrollState= new Boolean(jc.getAutoscrolls());
-					jc.setAutoscrolls(false);
-				}
-			}
-
-			//System.out.println("DragSourceDragEvent-dragEnter");
-	//		dsde.getDragSourceContext().
-		}
-		/**
-		 * Called as the hotspot exits a platform dependent drop site.
-		 */
-		public void dragExit(DragSourceEvent dse) {
-		}
-		/**
-		 * Called as the hotspot moves over a platform dependent drop site.
-		 */
-		public void dragOver(DragSourceDragEvent dsde) {
-			//System.out.println("DragSourceDragEvent-dragOver");
-		}
-		/**
-		 * Called when the user has modified the drop gesture.
-		 */
-		public void dropActionChanged(DragSourceDragEvent dsde) {
-		}
+	protected DragSourceListener createDragSourceListener(){
+		return new JHDDragSourceListener(editor(),view());
 	}
 }
 	/**
