+++ jhotdraw6/src/org/jhotdraw/standard/StandardDrawingView.java
@@ -1064,14 +1064,21 @@ public class StandardDrawingView
     }
 
 	protected DNDHelper createDNDHelper() {
-		return new DNDHelper () {
+		DNDHelper dndh = new DNDHelper () {
 				protected DrawingView view() {
 					return StandardDrawingView.this;
 				}
 				protected DrawingEditor editor() {
 					return StandardDrawingView.this.editor();
 				}
+				protected boolean isDropTarget(){
+					return true;
+				}
+				protected boolean isDragSource(){
+					return true;
+				}
 			};
+		return dndh;
 	}
 
 	protected DNDHelper getDNDHelper() {
@@ -1080,12 +1087,17 @@ public class StandardDrawingView
 		}
 		return dndh;
 	}
-
-	public boolean setDragSourceActive(boolean state) {
-		return getDNDHelper().setDragSourceActive(state);
+	public java.awt.dnd.DragSourceListener getDragSourceListener(){
+		return getDNDHelper().getDragSourceListener();
+	}
+	public void DNDInitialize(java.awt.dnd.DragGestureListener dgl){
+		getDNDHelper().initialize(dgl);
+	}
+	public void DNDDeinitialize() {
+		getDNDHelper().deinitialize();
 	}
 
-	public boolean setDropTargetActive(boolean state) {
-		return getDNDHelper().setDropTargetActive(state);
+	public void setDragSourceState(boolean state) {
+		getDNDHelper().setDragSourceState(state);
 	}
 }
